0x02 functions advance trails 15 and 16 2nd

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -1,20 +1,20 @@
 #include <stdio.h>
 /**
- * main - prints the first 98 Fibonacci numbers, starting with 1 and 2, 
+ * main - prints the first 98 Fibonacci numbers, starting with 1 and 2,
  * followed by a new line.
  * @void: no argument
  * Return: 0
  */
 int main(void)
 {
-	unsigned long pri = 1, sec = 2, sum = 0;
+	unsigned long long pri = 1, sec = 2, sum = 0;
 	int limit = 3;
 
 	printf("1, 2");
 	while (limit <= 98)
 	{
 		sum = pri + sec;
-		printf(", %lu", sum);
+		printf(", %llu", sum);
 		pri = sec;
 		sec = sum;
 		limit++;
